+++ react/components/ExtensionPoint.tsx
@@ -12,9 +12,11 @@ import { RenderContext } from './RenderContext'
 import TrackEventsWrapper from './TrackEventsWrapper'
 
 interface Props {
-  id: string,
-  params?: any,
-  query?: any,
+  id: string
+  params?: any
+  query?: any
+  treePath?: string
+  blockProps?: object
 }
 
 type ExtendedProps = Props & TreePathProps
@@ -25,10 +27,16 @@ interface State {
 
 class ExtensionPoint extends Component<ExtendedProps, State> {
   public static propTypes = {
+    blockProps: PropTypes.object,
     children: PropTypes.node,
+    id: PropTypes.string,
     params: PropTypes.object,
     query: PropTypes.object,
-    treePath: PropTypes.string.isRequired,
+    treePath: PropTypes.string,
+  }
+
+  public static defaultProps = {
+    blockProps: {},
   }
 
   public static childContextTypes = {
@@ -81,15 +89,41 @@ class ExtensionPoint extends Component<ExtendedProps, State> {
 
   private getExtensionPointComponent = (runtime: RenderContext) => {
     const { newTreePath } = this.state
-    const { children, params, query, id, treePath, ...parentProps } = this.props
+    const {
+      children,
+      params,
+      query,
+      id,
+      treePath,
+      blockProps,
+      ...parentProps
+    } = this.props
+
     const extension = runtime.extensions && runtime.extensions[newTreePath]
-    const { component = null, after = [], around = [], before = [], content = {}, props: extensionProps = {}, track = [] } = extension || {}
+    const {
+      component = null,
+      after = [],
+      around = [],
+      before = [],
+      content = {},
+      props: extensionProps = {},
+      track = []
+    } = extension || {}
 
     this.component = component
 
     const props = reduce(mergeDeepRight, {}, [
+      /** Extra props passed to the ExtensionPoint component
+       * e.g. <ExtensionPoint foo="bar" />
+       */
       parentProps,
+      /** Props that are read from runtime.extensions, that come from the blocks files
+       */
       extensionProps,
+      /** Props from the blockProps prop, used when the user wants to prevent overriding
+       * the native ExtensionPoint props (such as `id`)
+       */
+      blockProps,
       content,
       { params, query },
     ])
@@ -109,13 +143,7 @@ class ExtensionPoint extends Component<ExtendedProps, State> {
     const isCompositionChildren = extension && extension.composition === 'children'
 
     const componentChildren = (isCompositionChildren && extension.blocks) ?
-      extension.blocks.map((child, i) =>
-        <ExtensionPoint
-          key={i}
-          id={child.extensionPointId}
-          treePath={newTreePath}
-        />
-      ) : children
+      this.getChildExtensions(runtime, newTreePath) : children
 
     return component
       ? this.withOuterExtensions(
@@ -137,6 +165,47 @@ class ExtensionPoint extends Component<ExtendedProps, State> {
           )
         )
       : loading
+  }
+
+  private getChildExtensions(runtime: RenderContext, treePath: string) {
+    const extension = runtime.extensions && runtime.extensions[treePath]
+
+    if (!extension || !extension.blocks) {
+      return
+    }
+
+    return extension.blocks.map((child, i) => {
+      const childTreePath = ExtensionPoint.mountTreePath(child.extensionPointId, treePath)
+      const childExtension = runtime.extensions && runtime.extensions[childTreePath]
+      const childProps = childExtension ? childExtension.props : {}
+
+      /* This ExtensionPointWrapper thing is done so the user can read
+        * the props that were passed through the blocks.json file to
+        * its children in a standard, React-ish way; that is:
+        * `React.Children.map(children, child => child.props)`
+        * 
+        * The problem was, if the user passed a prop that conflicted with
+        * ExtensionPoint props (most notabily, `id`), just destructuring
+        * the `childProps` over ExtensionPoint would override the 
+        * ExtensionPoint props, which would break the rendering.
+        * (or vice versa, which would cause wrong values being read by
+        * the user component). 
+      */
+      const ExtensionPointWrapper = (blockProps: object) => (
+        <ExtensionPoint
+          id={child.extensionPointId}
+          treePath={treePath}
+          blockProps={blockProps}
+        />
+      )
+
+      return (
+        <ExtensionPointWrapper
+          key={i}
+          {...childProps}
+        />
+      )
+    })
   }
 
   private withOuterExtensions(after: string[], around: string[], before: string[], treePath: string, props: any, element: JSX.Element) {