+++ react/core/main.tsx
@@ -181,6 +181,10 @@ const TreePathContextConsumer = TreePathContext.Consumer
 
 export {
   ExtensionContainer,
+  /** Block is the preferred nomenclature now, ExtensionPoint is kept for
+   * backwards compatibility
+   */
+  ExtensionPoint as Block,
   ExtensionPoint,
   LayoutContainer,
   LegacyExtensionContainer,
@@ -194,6 +198,9 @@ export {
   start,
   withHMR,
   withRuntimeContext,
+  ChildBlock,
+  useChildBlock,
+  // These unstable APIs should be deprecated shortly
   ChildBlock as Unstable__ChildBlock,
   useChildBlock as useChildBlock__unstable,
   useRuntime,