+++ CHANGELOG.md
@@ -7,6 +7,17 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [8.18.0] - 2019-04-10
+### Added
+- Export ChildBlock and useChildBlock APIs, without unstable flag.
+- ChildBlock now returns the props of the block, inserted via the `blocks.json` file.
+- Send props inserted via the `blocks.json` file to the `props` object of each child (when `composition` is set to `children`).
+- Export Block component, which is an alias of the ExtensionPoint component, and is now the preferred nomenclature.
+
+### Deprecated
+- Unstable__ChildBlock and useChildBlock__unstable.
+- ExtensionPoint component--prefer using the Block component, which has exactly the same API and functionality.
+
 ## [8.17.4] - 2019-04-05
 ### Changed
 - Update vtex-tachyons to 3.1.0